requirements.txt
altgraph==0.17.4
colorama==0.4.6
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pefile==2023.2.7
pillow==10.2.0
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
PyQt6-Qt6==6.6.2
PyQt6-sip==13.6.0
PySide6==6.6.2
PySide6_Addons==6.6.2
PySide6_Essentials==6.6.2
pywin32-ctypes==0.2.2
shiboken6==6.6.2
tqdm==4.66.2
